Rename config.py to interfaces.py
